+++ training/src/anemoi/training/losses/huber.py
@@ -13,19 +13,18 @@ import logging
 
 import torch
 
-from anemoi.training.losses.weightedloss import BaseWeightedLoss
+from anemoi.training.losses.base import FunctionalLoss
 
 LOGGER = logging.getLogger(__name__)
 
 
-class WeightedHuberLoss(BaseWeightedLoss):
-    """Node-weighted Huber loss."""
+class HuberLoss(FunctionalLoss):
+    """Huber loss."""
 
-    name = "whuber"
+    name: str = "huber"
 
     def __init__(
         self,
-        node_weights: torch.Tensor,
         delta: float = 1.0,
         ignore_nans: bool = False,
         **kwargs,
@@ -36,21 +35,15 @@ class WeightedHuberLoss(BaseWeightedLoss):
 
         Parameters
         ----------
-        node_weights : torch.Tensor of shape (N, )
-            Weight of each node in the loss function
         delta : float, optional
             Threshold for Huber loss, by default 1.0
         ignore_nans : bool, optional
             Allow nans in the loss and apply methods ignoring nans for measuring the loss, by default False
         """
-        super().__init__(
-            node_weights=node_weights,
-            ignore_nans=ignore_nans,
-            **kwargs,
-        )
+        super().__init__(ignore_nans=ignore_nans, **kwargs)
         self.delta = delta
 
-    def huber(self, pred: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
+    def calculate_difference(self, pred: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
         """Calculate the Huber loss.
 
         Parameters
@@ -67,38 +60,3 @@ class WeightedHuberLoss(BaseWeightedLoss):
         """
         diff = torch.abs(pred - target)
         return torch.where(diff < self.delta, 0.5 * torch.square(diff), self.delta * (diff - 0.5 * self.delta))
-
-    def forward(
-        self,
-        pred: torch.Tensor,
-        target: torch.Tensor,
-        squash: bool = True,
-        scalar_indices: tuple[int, ...] | None = None,
-        without_scalars: list[str] | list[int] | None = None,
-    ) -> torch.Tensor:
-        """Calculates the lat-weighted Huber loss.
-
-        Parameters
-        ----------
-        pred : torch.Tensor
-            Prediction tensor, shape (bs, ensemble, lat*lon, n_outputs)
-        target : torch.Tensor
-            Target tensor, shape (bs, ensemble, lat*lon, n_outputs)
-        squash : bool, optional
-            Average last dimension, by default True
-        scalar_indices: tuple[int,...], optional
-            Indices to subset the calculated scalar with, by default None
-        without_scalars: list[str] | list[int] | None, optional
-            list of scalars to exclude from scaling. Can be list of names or dimensions to exclude.
-            By default None
-
-        Returns
-        -------
-        torch.Tensor
-            Weighted Huber loss
-        """
-        out = self.huber(pred, target)
-
-        out = self.scale(out, scalar_indices, without_scalars=without_scalars)
-
-        return self.scale_by_node_weights(out, squash)