+++ training/src/anemoi/training/schemas/graphs/node_attributes_schemas.py
@@ -14,8 +14,9 @@ import logging
 from typing import Literal
 from typing import Union
 
-from pydantic import BaseModel
 from pydantic import Field
+
+from anemoi.training.schemas.utils import BaseModel
 
 LOGGER = logging.getLogger(__name__)